FOLSPRINGB-1: Move tenant init logic to a service class.
