+++ src/main/java/org/folio/spring/config/DefaultExceptionHandler.java
@@ -1,5 +1,6 @@
 package org.folio.spring.config;
 
+import org.folio.spring.exception.NotFoundException;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.MediaType;
@@ -32,4 +33,8 @@ public class DefaultExceptionHandler {
     return ResponseEntity.badRequest().contentType(MediaType.TEXT_PLAIN).body(e.getMessage());
   }
 
+  @ExceptionHandler(NotFoundException.class)
+  public ResponseEntity<String> handleNotFoundException(NotFoundException ex) {
+    return ResponseEntity.status(404).body(ex.getMessage());
+  }
 }
