+++ src/main/java/org/folio/spring/config/FolioSpringConfiguration.java
@@ -13,7 +13,8 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan({"org.folio.spring.controller", "org.folio.spring.config", "org.folio.spring.filter"})
+@ComponentScan({"org.folio.spring.controller", "org.folio.spring.config",
+  "org.folio.spring.filter", "org.folio.spring.service"})
 public class FolioSpringConfiguration {
 
   @Bean
