+++ src/main/java/org/folio/spring/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package org.folio.spring.exception;
+
+public class NotFoundException extends RuntimeException {
+  public NotFoundException(String message) {
+    super(message);
+  }
+}
